Исправляет ошибку с закрытием попапа + переименовывает компоненты со списками и отделяет контейнер films в отдельный компонент
