+++ src/main.js
@@ -2,15 +2,16 @@ import HeaderProfileView from './view/header-profile';
 import SiteMenuView from './view/site-menu';
 import FooterStatisticView from './view/footer-statistics';
 import SortView from './view/sort';
-import FilmsListView from './view/films-list';
+import FilmListView from './view/film-list';
 import NoFilmView from './view/no-film';
 import FilmCardView from './view/film-card';
 import ShowMoreButtonView from './view/show-more-btn';
-import FilmsListExtraView from './view/films-list-extra';
+import FilmListExtraView from './view/film-list-extra';
 import FilmDetailsView from './view/film-details';
 import { generateFilm } from './mock/film';
 import { allComments } from './mock/comment';
 import { generateFilter } from './mock/filter';
+import FilmsView from './view/films';
 import { render, remove, RenderPosition } from './utils/render';
 
 const FILMS_COUNT = 12;
@@ -29,16 +30,18 @@ const mainElement = document.querySelector('.main');
 const footerStats = document.querySelector('.footer__statistics');
 
 render(headerElement, new HeaderProfileView(), RenderPosition.BEFOREEND);
+render(footerStats, new FooterStatisticView(filmItems.length), RenderPosition.BEFOREEND);
 render(mainElement, new SiteMenuView(filters), RenderPosition.BEFOREEND);
 
-if (filmItems.length === 0) {
-  render(mainElement, new NoFilmView(), RenderPosition.BEFOREEND);
-} else {
-  render(mainElement, new SortView(), RenderPosition.BEFOREEND);
-  render(mainElement, new FilmsListView(), RenderPosition.BEFOREEND);
-}
+const filmsComponent = new FilmsView();
+render(mainElement, filmsComponent, RenderPosition.BEFOREEND);
 
-render(footerStats, new FooterStatisticView(filmItems.length), RenderPosition.BEFOREEND);
+if (filmItems.length === 0) {
+  render(filmsComponent, new NoFilmView(), RenderPosition.BEFOREEND);
+} else {
+  render(filmsComponent, new SortView(), RenderPosition.BEFOREEND);
+  render(filmsComponent, new FilmListView(), RenderPosition.BEFOREEND);
+}
 
 const renderFilmCard = (container, film) => {
   const filmCard = new FilmCardView(film);
@@ -48,6 +51,7 @@ const renderFilmCard = (container, film) => {
     render(document.body, filmDetails, RenderPosition.BEFOREEND);
     document.body.classList.add('hide-overflow');
     document.addEventListener('keydown', onEscKeyDown);
+    filmDetails.setCloseButtonClickHandler(hideDetails);
   };
 
   const hideDetails = () => {
@@ -64,12 +68,11 @@ const renderFilmCard = (container, film) => {
   };
 
   filmCard.setFilmCardClickHandler(showDetails);
-  filmDetails.setCloseButtonClickHandler(hideDetails);
   render(container, filmCard, RenderPosition.BEFOREEND);
 };
 
 const filmsElement = mainElement.querySelector('.films');
-const filmsListElement = filmsElement.querySelector('.films-list');
+const filmListElement = filmsElement.querySelector('.films-list');
 const filmListContainerElement = filmsElement.querySelector('.films-list__container');
 
 for (let i = 0; i < Math.min(filmItems.length, FILMS_COUNT_PER_STEP); i++) {
@@ -80,7 +83,7 @@ if (filmItems.length > FILMS_COUNT_PER_STEP) {
   let renderedFilmCount = FILMS_COUNT_PER_STEP;
   const showMoreButton = new ShowMoreButtonView();
 
-  render(filmsListElement, showMoreButton, RenderPosition.BEFOREEND);
+  render(filmListElement, showMoreButton, RenderPosition.BEFOREEND);
 
   const showMoreCards = () => {
     filmItems
@@ -97,8 +100,8 @@ if (filmItems.length > FILMS_COUNT_PER_STEP) {
 }
 
 if (filmItems.length > 0) {
-  render(filmsElement, new FilmsListExtraView('Top rated'), RenderPosition.BEFOREEND);
-  render(filmsElement, new FilmsListExtraView('Most commented'), RenderPosition.BEFOREEND);
+  render(filmsComponent, new FilmListExtraView('Top rated'), RenderPosition.BEFOREEND);
+  render(filmsComponent, new FilmListExtraView('Most commented'), RenderPosition.BEFOREEND);
   const topRatedFilmsContainer = filmsElement.querySelector('#top-rated .films-list__container');
   const mostCommentedFilmsContainer = filmsElement.querySelector('#most-commented .films-list__container');
   for (let i = 0; i < EXTRA_LIST_FILMS_COUNT; i++) {
