+++ src/view/films.js
@@ -1,0 +1,9 @@
+import AbstractView from './abstract';
+
+class Films extends AbstractView {
+  getTemplate() {
+    return '<section class="films"></section>';
+  }
+}
+
+export { Films as default };
